Alleviate constant container restarts during dev
Previously we used COPY to put our code into the container when building
it. That meant every little code change required the container to be
rebuilt. Now the whole workspace is shared to the container as a
volume. Still need to restart Ruby when .rb files have changed, but it's
a LOT faster than running bundle install all over again every time.