Use new model instance for each request
The app was using up a lot of memory. That's probably because we
stored the DB connection in a global variable in which case Sequel
keeps the connection open by default.

This commit re-architects the code to make a new DB connection for
each client request. No connection is kept in the background.
Hopefully this will dramatically reduces memory usage when idle.

Details on Sequel GC from https://stackoverflow.com/a/16072796

    Note that if you do not pass a block to Sequel.connect, Sequel
    will automatically retain a reference to the object in the
    Sequel::DATABASES array. So calling Sequel.connect mul tiple times
    (say once per request), can result in a memory leak. For any
    application where database access is needed for a long period of
    time, it's best to store the result of Sequel.connection in a
    constant, as recommended above.